refactor(cli/test): moved core e2e tests to the core folder
